Add dev-lang/fsharp-4.0.0.4-r1 ebuild.
Changes wrt to fsharp-4.0.0.4 in portage:
- Added pkg_setup() function to allow writing to
  /usr/share/.mono/keypairs

If you get Nuget errors you need to run
`cert-sync /etc/ssl/certs/ca-certificates.crt` which I think should be
done by the dev-lang/mono ebuild (but is not).

Also note that this ebuild uses the network to update some packages
which is a bad thing. Possible solution from OpenSUSE Build service, but
it requires NUnit installed (which is not in portage, only in dotnet
overlay).